Lithium

Documentation for Lithium.

Lithium.coerce_scitypes

Lithium.compute_alda_b_score

Lithium.compute_alda_total_score

Lithium.merge_mania_hypomania_ageofonset

Lithium.narrow_comorbidity_definition

Lithium.normalize_family_history

Lithium.probands_only

Lithium.prune_missingness

Lithium.remove_columns

Lithium.remove_excluded_variables

Lithium.remove_ids

Lithium.remove_incomplete_targets

Lithium.replace_centre_names

Lithium.coerce_scitypes — Method

coerce_scitypes(data)

Function to coerce to appropriate scientific types, such that it can be piped.

Lithium.compute_alda_b_score — Function

compute_alda_b_score(data[, med])

Sum over Alda B scores for med ∈ ["li", "vpa", "ltg"]

Lithium.compute_alda_total_score — Function

compute_alda_total_score(data[, med])

Sum over Alda total scores.

Lithium.merge_mania_hypomania_ageofonset — Method

merge_mania_hypomania_ageofonset(data)

Combines the ages of mania and hypomania onset into a single variable where the minimum is used as the age of onset for the union of these phases

Lithium.narrow_comorbidity_definition — Method

narrow_comorbidity_definition(data)

For the comorbidity fields "social_anxiety", "panic_disorder", "generalized_anxiety", "ocd", "substance_abuse", "adhd","learning_disability", "primary_insomnia", "personality_disorder" we set a binary indicator as positive only if they met full criteria.

Lithium.normalize_family_history — Method

normalize_family_history(data)

Normalizes family history counts to [0,1]. If no first degree relatives, value is 0.

Lithium.probands_only — Method

probands_only(data)

Excludes any relatives from the dataset.

Lithium.prune_missingness — Function

prune_missingness(data, p)

Function that identifies which features have less than p (proportion) of values missing.

Lithium.remove_columns — Function

remove_columns(data, colregex[, fuzzy])

Selects all columns in data except those matching the colregex pattern. If fuzzy=false then only columns matching colregex exactly will be removed.

Lithium.remove_excluded_variables — Method

narrow_comorbidity_definition(data)

Removes variables that are excluded from the analysis based on being (A) after medication initiation, or deemed irrelevant a priori. These variables are "dob", "assessmentdate", "mostrecentupdate", "affectedstatus", "gaflastassessment", "maritalstatus", "ses", "primarydiagnosis", "nepisodesonli", "fatherethnicity", "motherethnicity", "diagnosis", "totallifetimeepisodes", "nseconddegreerelcompletedsuicide", "nseconddegreerelsuicideattempt", "nsuicideattempts", "lifetimemanias", "lifetimedepressions", "lifetimemixed", "lifetimemultiphasic", "lifetimepsychosis", and "lifetimehxsuicide_ideation".

Lithium.remove_ids — Method

remove_ids(data)

Removes family and person ID's.

Lithium.remove_incomplete_targets — Method

remove_incomplete_targets(data, regex_str)

Includes only those rows that have complete records for the specified column

Lithium.replace_centre_names — Method

replace_centre_names(data)

Replace centre names with their acronyms.